# [ResourceItem] Remove children from the equality check in shouldComponentUpdate