@jineshshah36 2020-05-15
Hi, we merged this change it broke a bunch of our `ResourceItem` components. I have put together a codesandbox [here](https://codesandbox.io/s/flamboyant-river-8v1zm?file=/App.js) which exemplifies the problem.

Essentially, if you attempt to update the children of the resource item asynchronously without making any additional prop changes, it will never re-render. In the example I included, the author is "fetched" asynchronously, and when we attempt to re-render with the author info, it doesn't.

I believe excluding children is simply a cure for the symptom, but the real problem is that the code is using deep equals instead of shallow equals. Is there a reason for the deep equals? Shouldn't all props be immutable, therefore not need a deep equals comparison?